backend/requirements.txt
flask==3.1.2
flask-cors==6.0.1
python-dotenv==1.2.1
pandas==2.3.3
numpy==2.3.5
scikit-learn==1.8.0
joblib==1.5.2
xgboost==3.1.2
lightgbm==4.6.0
catboost==1.2.8
mlflow==3.7.0
alembic==1.17.2
fastapi==0.124.4
uvicorn==0.38.0
SQLAlchemy==2.0.45
seaborn==0.13.2
matplotlib==3.10.8
PyYAML==6.0.3
requests==2.32.5
plotly==6.5.0

